Moved kcp test files to new, "cards", directory
